Disable `KUMI_BUILD_TEST` in `FetchContent` docs.
`copacabana` defaults to `${CMAKE_SOURCE_DIR}/test` when searching
for source files in `copa_glob_unit`, this will refer to the
top-level directory, ie the users project directory. The paths
of the sources will then be `../path-to-build/_deps/kumi-src/...`
and so on, in addition it might also match sources in the user
project.

Later `copa_make_unit` does `add_executable` with the same sources,
but these sources will be interpreter relative to the source
directory of the target, not `${CMAKE_SOURCE_DIR}/test` and then
no longer refer to an existing file.

Likely `copacabana` should refer to `${PROJECT_SOURCE_DIR}` instead
of `${CMAKE_SOURCE_DIR}` so that it is relative to the root of the
most recently declared projet, ie the one that is calling `copacabana`
functions. Until then disable `KUMI_BUILD_TEST` in the docs.